+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
+### Added
+- Logging when importing dynamic avatars
 
 ## [4.0.0-alpha9] 2018-01-23
 ### Added